Remove agent-sdk-ts runtime/type import cycles (oh-tab-zhwi.6.1)
