+++ packages/agent-sdk-ts/src/sdk/runtime/stateTypes.ts
@@ -1,0 +1,5 @@
+export interface AgentState {
+  status: string;
+  iteration: number;
+  values: Record<string, unknown>;
+}
